- add api route - add graphql - add graphql schema - update sequalize dir to db/ - add demo route for users and products
