+++ server/.sequelizerc
@@ -1,0 +1,8 @@
+const path = require('path')
+
+module.exports = {
+  'config':          path.resolve('sequelize.config.json'),
+  'migrations-path': path.resolve('db', 'migrations'),
+  'models-path':     path.resolve('db', 'models'),
+  'seeders-path':    path.resolve('db', 'seeders'),
+}
