+++ server/db/models/user.js
@@ -1,0 +1,16 @@
+"use strict";
+module.exports = (sequelize, DataTypes) => {
+    const User = sequelize.define(
+        "users",
+        {
+            firstName: DataTypes.STRING,
+            lastName: DataTypes.STRING,
+            email: DataTypes.STRING
+        },
+        {}
+    );
+    User.associate = function(models) {
+        // associations can be defined here
+    };
+    return User;
+};
