+++ server/graphql/schema.js
@@ -1,0 +1,34 @@
+const graphql = require("graphql");
+const _ = require("lodash");
+const { GraphQLObjectType, GraphQLString, GraphQLSchema } = graphql;
+
+const UserType = new GraphQLObjectType({
+    name: "User",
+    fields: {
+        id: { type: GraphQLString },
+        firstName: { type: GraphQLString },
+        lastName: { type: GraphQLString }
+    }
+});
+
+const users = [
+    { id: "23", firstName: "bill", age: 20 },
+    { id: "43", firstName: "ted", age: 20 }
+];
+
+const RootQuery = new GraphQLObjectType({
+    name: "RootQueryType",
+    fields: {
+        user: {
+            type: UserType,
+            args: { id: { type: GraphQLString } },
+            resolve(parentValue, args) {
+                return _.find(users, { id: args.id });
+            }
+        }
+    }
+});
+
+module.exports = new GraphQLSchema({
+    query: RootQuery
+});
