+++ server/package.json
@@ -5,7 +5,10 @@
     "main": "index.js",
     "dependencies": {
         "express": "^4.16.4",
+        "express-graphql": "^0.8.0",
+        "graphql": "^14.2.1",
         "helmet": "^3.15.0",
+        "lodash": "^4.17.11",
         "mysql2": "^1.6.5",
         "sequelize": "^5.5.0"
     },
