+++ server/server.js
@@ -1,12 +1,21 @@
 const express = require("express");
 const helmet = require("helmet");
 const app = express();
-const hello = require("./hello");
+const expressGraphQL = require("express-graphql");
+const schema = require("./graphql/schema");
+const api = require("./api");
 
 app.use(helmet());
 const port = 3030;
 
 app.get("/", (req, res) => res.send("Hello World!"));
-app.use("/hello", hello);
+app.use(
+    "/graphql",
+    expressGraphQL({
+        schema,
+        graphiql: true
+    })
+);
+app.use("/api", api);
 
 app.listen(port, () => console.log(`Example app listening on port ${port}!`));
